requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
aiohttp=3.8.1=py38h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=8.0.0=py38h60b952e_0
async-timeout=4.0.2=py38h06a4308_0
attrs=21.4.0=pyhd3eb1b0_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
blas=1.0=mkl
boost-cpp=1.73.0=h7f8727e_12
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
cryptography=38.0.1=py38h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.7.99=0
cuda-gdb=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
datasets=2.6.1=py38h06a4308_0
dill=0.3.6=py38h06a4308_0
ffmpeg=4.3=hf484d3e_0
filelock=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.2.0=py38h7f8727e_0
fsspec=2022.10.0=py38h06a4308_0
gds-tools=1.4.0.31=0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h7b6447c_0
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_0
huggingface_hub=0.10.1=py38h06a4308_0
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h7f8727e_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.73.0=h28710b8_12
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcurl=7.85.0=h91b91d3_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=hcc01f38_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.3=h295c915_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=py38h5eee18b_0
multiprocess=0.70.14=py38h06a4308_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
nltk=3.7=pyhd3eb1b0_0
nsight-compute=2022.3.0.22=0
numexpr=2.8.3=py38h807cd23_0
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py38h6a678d5_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
pyarrow=8.0.0=py38h992f0b0_0
pycparser=2.21=pyhd3eb1b0_0
pymongo=3.12.0=py38h295c915_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=haa1d7c7_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xxhash=2.0.2=py38h5eee18b_1
pytorch=1.13.0=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py38h06a4308_0
pyyaml=6.0=py38h7f8727e_1
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py38h5eee18b_0
requests=2.28.1=py38h06a4308_0
responses=0.13.3=pyhd3eb1b0_0
setuptools=65.5.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.39.3=h5082296_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py38h3dcd8bd_1
torchaudio=0.13.0=py38_cu117
torchmetrics=0.10.1=pyhd8ed1ab_0
torchvision=0.14.0=py38_cu117
tqdm=4.64.1=py38h06a4308_0
transformers=4.24.0=py38h06a4308_0
typing-extensions=4.3.0=py38h06a4308_0
typing_extensions=4.3.0=py38h06a4308_0
urllib3=1.26.12=py38h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=0.8.0=h7f8727e_3
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py38h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
